refactored utils
